requirements.txt
qrcode>=7.4.2
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.2
ttkthemes>=3.2.2
pycryptodome>=3.18.0
pyzbar>=0.1.9
